Change hooks to typescript.
